Conflicts:
	motion/Observer.cpp
	noggin/NogginStructs.h
	vision/Ball.cpp
	vision/Field.cpp
	vision/VisualCorner.h
	vision/VisualLine.cpp

+++ noggin/Noggin.cpp
@@ -431,7 +431,7 @@ void Noggin::updateLocalization()
         if (!(n.ballX == 0.0 && n.ballY == 0.0) &&
             !(gc->gameState() == STATE_INITIAL ||
               gc->gameState() == STATE_FINISHED)) {
-            m.distance = hypot(loc->getXEst() - n.ballX,
+            m.distance = hypotf(loc->getXEst() - n.ballX,
                                loc->getYEst() - n.ballY);
             m.bearing = subPIAngle(atan2(n.ballY - loc->getYEst(),
                                          n.ballX - loc->getXEst()) -
